move main method to BankAccountTest
