+++ src/main/java/io/codelex/classesandobjects/practice/BankAccount.java
@@ -7,13 +7,6 @@ public class BankAccount {
     public String name;
     public double balance;
 
-    public static void main(String[] args) {
-        BankAccount ben = new BankAccount("Benson", 17.25);
-        System.out.println(ben);
-
-        BankAccount benNegative = new BankAccount("Benson", -17.25);
-        System.out.println(benNegative);
-    }
 
     public BankAccount(String name, double balance) {
         this.name = name;
